+++ src/main/java/project/backend/common/error/exception/UserNotAuthenticatedException.java
@@ -2,10 +2,6 @@ package project.backend.common.error.exception;
 
 public class UserNotAuthenticatedException extends RuntimeException {
 
-  public UserNotAuthenticatedException() {
-    super("인증된 유저 정보를 찾을 수 없습니다.");
-  }
-
   public UserNotAuthenticatedException(String message) {
     super(message);
   }
